requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
djangorestframework==3.12.1
importlib-metadata==1.7.0
Pillow==8.0.1
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
